+++ test/gallery/gallery_screen_test.dart
@@ -27,8 +27,8 @@ import 'package:wger/providers/gallery.dart';
 import 'package:wger/screens/form_screen.dart';
 import 'package:wger/widgets/gallery/overview.dart';
 
-import '../test_data/gallery.dart';
-import 'gallery_screen_test.mocks.dart';
+import './gallery_screen_test.mocks.dart';
+import '../../test_data/gallery.dart';
 
 @GenerateMocks([GalleryProvider])
 void main() {
