+++ test/nutrition/nutritional_meal_form_test.dart
@@ -19,36 +19,36 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
 import 'package:wger/helpers/consts.dart';
+import 'package:wger/helpers/json.dart';
+import 'package:wger/models/nutrition/meal.dart';
 import 'package:wger/models/nutrition/nutritional_plan.dart';
 import 'package:wger/providers/nutrition.dart';
 import 'package:wger/screens/nutritional_plan_screen.dart';
 import 'package:wger/widgets/nutrition/forms.dart';
 
-import 'nutritional_plan_form_test.mocks.dart';
+import './nutritional_plan_form_test.mocks.dart';
+import '../../test_data/nutritional_plans.dart';
 
-@GenerateMocks([NutritionPlansProvider])
 void main() {
   var mockNutrition = MockNutritionPlansProvider();
 
-  final plan1 = NutritionalPlan(
-    id: 1,
-    creationDate: DateTime(2021, 1, 1),
-    description: 'test plan 1',
-  );
-  final plan2 = NutritionalPlan.empty();
+  var plan1 = NutritionalPlan.empty();
+  var meal1 = Meal();
+  final meal2 = Meal();
 
-  when(mockNutrition.editPlan(any)).thenAnswer((_) => Future.value(plan1));
-  when(mockNutrition.addPlan(any)).thenAnswer((_) => Future.value(plan2));
+  when(mockNutrition.editMeal(any)).thenAnswer((_) => Future.value(Meal()));
+  when(mockNutrition.addMeal(any, any)).thenAnswer((_) => Future.value(Meal()));
 
   setUp(() {
+    plan1 = getNutritionalPlan();
+    meal1 = plan1.meals.first;
     mockNutrition = MockNutritionPlansProvider();
   });
 
-  Widget createHomeScreen(NutritionalPlan plan, {locale = 'en'}) {
+  Widget createHomeScreen(Meal meal, {locale = 'en'}) {
     final key = GlobalKey<NavigatorState>();
 
     return ChangeNotifierProvider<NutritionPlansProvider>(
@@ -59,7 +59,7 @@ void main() {
         supportedLocales: AppLocalizations.supportedLocales,
         navigatorKey: key,
         home: Scaffold(
-          body: PlanForm(plan),
+          body: MealForm(1, meal),
         ),
         routes: {
           NutritionalPlanScreen.routeName: (ctx) => NutritionalPlanScreen(),
@@ -68,59 +68,59 @@ void main() {
     );
   }
 
-  testWidgets('Test the widgets on the nutritional plan form', (WidgetTester tester) async {
-    await tester.pumpWidget(createHomeScreen(plan1));
+  testWidgets('Test the widgets on the meal form', (WidgetTester tester) async {
+    await tester.pumpWidget(createHomeScreen(meal1));
     await tester.pumpAndSettle();
 
-    expect(find.byType(TextFormField), findsOneWidget);
+    expect(find.byType(TextFormField), findsNWidgets(2));
     expect(find.byType(ElevatedButton), findsOneWidget);
     expect(find.byKey(const Key(SUBMIT_BUTTON_KEY_NAME)), findsOneWidget);
   });
 
-  testWidgets('Test editing an existing nutritional plan', (WidgetTester tester) async {
-    await tester.pumpWidget(createHomeScreen(plan1));
+  testWidgets('Test editing an existing meal', (WidgetTester tester) async {
+    await tester.pumpWidget(createHomeScreen(meal1));
     await tester.pumpAndSettle();
 
     expect(
-      find.text('test plan 1'),
+      find.text('17:00'),
       findsOneWidget,
-      reason: 'Description of existing nutritional plan is filled in',
+      reason: 'Time of existing meal is filled in',
     );
-    await tester.enterText(find.byKey(const Key('field-description')), 'New description');
+
+    expect(
+      find.text('Initial Name 1'),
+      findsOneWidget,
+      reason: 'Time of existing meal is filled in',
+    );
+
+    await tester.enterText(find.byKey(const Key('field-time')), '12:34');
+    await tester.enterText(find.byKey(const Key('field-name')), 'test meal');
     await tester.tap(find.byKey(const Key(SUBMIT_BUTTON_KEY_NAME)));
 
     // Correct method was called
-    verify(mockNutrition.editPlan(any));
-    verifyNever(mockNutrition.addPlan(any));
-
-    // TODO(x): edit calls Navigator.pop(), since the form can only be reached from the
-    //       detail page. The test needs to add the detail page to the stack so that
-    //       this can be checked.
-    // https://stackoverflow.com/questions/50704647/how-to-test-navigation-via-navigator-in-flutter
-
-    // Detail page
-    //await tester.pumpAndSettle();
-    //expect(
-    // find.text(('New description')),
-    //findsOneWidget,
-    //reason: 'Nutritional plan detail page',
-    //);
+    verify(mockNutrition.editMeal(any));
+    verifyNever(mockNutrition.addMeal(any, any));
   });
 
   testWidgets('Test creating a new nutritional plan', (WidgetTester tester) async {
-    await tester.pumpWidget(createHomeScreen(plan2));
+    await tester.pumpWidget(createHomeScreen(meal2));
     await tester.pumpAndSettle();
 
-    expect(find.text(''), findsOneWidget, reason: 'New nutritional plan has no description');
-    await tester.enterText(find.byKey(const Key('field-description')), 'New cool plan');
+    expect(
+      find.text(timeToString(TimeOfDay.now())!),
+      findsOneWidget,
+      reason: 'Current time is filled in',
+    );
+
+    await tester.enterText(find.byKey(const Key('field-time')), '08:00');
+    await tester.enterText(find.byKey(const Key('field-name')), 'test meal');
     await tester.tap(find.byKey(const Key(SUBMIT_BUTTON_KEY_NAME)));
 
     // Correct method was called
-    verifyNever(mockNutrition.editPlan(any));
-    verify(mockNutrition.addPlan(any));
+    verifyNever(mockNutrition.editMeal(any));
+    verify(mockNutrition.addMeal(any, any));
 
     // Detail page
-    await tester.pumpAndSettle();
-    expect(find.text('New cool plan'), findsOneWidget, reason: 'Nutritional plan detail page');
+    // ...
   });
 }
