+++ test/workout/workout_plan_screen_test.dart
@@ -23,9 +23,9 @@ import 'package:provider/provider.dart';
 import 'package:wger/providers/workout_plans.dart';
 import 'package:wger/screens/workout_plan_screen.dart';
 
-import '../test_data/workouts.dart';
-import 'base_provider_test.mocks.dart';
-import 'utils.dart';
+import '../../test_data/workouts.dart';
+import '../other/base_provider_test.mocks.dart';
+import '../utils.dart';
 
 void main() {
   Widget createHomeScreen({locale = 'en'}) {
